+++ library/std/src/sys/wasm/mod.rs
@@ -50,22 +50,23 @@ pub mod time;
 cfg_if::cfg_if! {
     if #[cfg(target_feature = "atomics")] {
         #[path = "atomics/condvar.rs"]
-        pub mod condvar;
+        mod condvar;
         #[path = "atomics/mutex.rs"]
-        pub mod mutex;
+        mod mutex;
         #[path = "atomics/rwlock.rs"]
-        pub mod rwlock;
+        mod rwlock;
+        pub mod locks {
+            pub use super::condvar::*;
+            pub use super::mutex::*;
+            pub use super::rwlock::*;
+        }
         #[path = "atomics/futex.rs"]
         pub mod futex;
         #[path = "atomics/thread.rs"]
         pub mod thread;
     } else {
-        #[path = "../unsupported/condvar.rs"]
-        pub mod condvar;
-        #[path = "../unsupported/mutex.rs"]
-        pub mod mutex;
-        #[path = "../unsupported/rwlock.rs"]
-        pub mod rwlock;
+        #[path = "../unsupported/locks/mod.rs"]
+        pub mod locks;
         #[path = "../unsupported/thread.rs"]
         pub mod thread;
     }
